api/requirements.txt
fastapi==0.116.1
SQLAlchemy==2.0.41
fastapi[standard]==0.116.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.10
passlib==1.7.4
dotenv==0.9.9